requirements.txt
pandas
requests
python-dotenv
python-docx
python-pptx
chardet
openpyxl
langchain-core
langchain-community
pypdf
pymupdf
langchain-text-splitters
langchain-huggingface
langchain-chroma
chromadb
rank_bm25
fastapi
uvicorn
pydantic
sentence_transformers